fix(rakki): Better UI
